Omdb.Page: Store min_key for every child.
This simplifies the structure of a node. Every child reference in a
node also holds the minimal key:

type child = {id: id; min_key key}

A node is now just a sequence of children.

Storing the min_key for the first child is not necessary when searching
in the tree. Previously it was not stored in the node.

When merging nodes, the minimal key of the first child might be
needed. Previously we were descending the tree to find the minimal
key. This may have caused many page reads. This is no longer necessary
as the minimal key is stored in the node.

This trades node density for merge efficiency and also data structure
simplicity. A good trade, IMHO.